Refactor sensor entities
